Move broadcast logic to concern
