+++ app/models/workspace_sub.rb
@@ -1,4 +1,6 @@
  class WorkspaceSub < ApplicationRecord
+  include Concerns::Broadcastable
+
   attr_accessor :skip_broadcast
 
   validates_presence_of :workspace_id, scope: :user_id
